# openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A simple pastebin service with analytics
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pastebin API
  version: "1.0"
paths:
  /analytics:
    get:
      description: Retrieve all analytics with pagination
      parameters:
        - description: Sort order
          in: query
          name: order
          schema:
            type: string
        - description: Limit number of results
          in: query
          name: limit
          schema:
            type: integer
        - description: Offset for pagination
          in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: List of analytics
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/models.Analytics"
                type: array
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to get analytics
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get all analytics
      tags:
        - analytics
  "/analytics/{id}":
    get:
      description: Retrieve specific analytics entry by its ID
      parameters:
        - description: Analytics ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Analytics"
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to get analytics
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get analytics by ID
      tags:
        - analytics
  /analytics/paste:
    get:
      description: Retrieve analytics for a specific paste
      parameters:
        - description: Paste ID
          in: query
          name: pasteID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Analytics"
        "400":
          description: Invalid paste ID
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to get analytics
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get analytics by paste ID
      tags:
        - analytics
  /analytics/user:
    get:
      description: Retrieve analytics for a specific user with pagination
      parameters:
        - description: User ID
          in: query
          name: userID
          required: true
          schema:
            type: string
        - description: Sort order
          in: query
          name: order
          schema:
            type: string
        - description: Limit number of results
          in: query
          name: limit
          schema:
            type: integer
        - description: Offset for pagination
          in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: List of analytics
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/models.Analytics"
                type: array
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to get analytics
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get analytics by user
      tags:
        - analytics
  /analytics/increment-view-count:
    get:
      description: Increment the view count for a specific paste
      parameters:
        - description: Paste ID
          in: query
          name: pasteID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: View count incremented successfully
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "400":
          description: Invalid paste ID
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to increment view count
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Increment view count
      tags:
        - analytics
  /analytics/url:
    get:
      description: Retrieve analytics for a specific URL
      parameters:
        - description: URL
          in: query
          name: url
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Analytics"
        "400":
          description: Invalid URL
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to get analytics
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get analytics by URL
      tags:
        - analytics
  /create-analytics:
    post:
      description: Create a new analytics entry for a paste
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/models.AnalyticsInput"
        description: Analytics data
        required: true
      responses:
        "201":
          description: Analytics created successfully
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to create analytics
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Create analytics entry
      tags:
        - analytics
...
servers:
  - url: //localhost:8080
components:
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
  schemas:
    models.Analytics:
      properties:
        created_at:
          type: string
        id:
          type: string
        paste_id:
          type: string
        updated_at:
          type: string
        url:
          type: string
        views:
          type: integer
      type: object
    models.AnalyticsInput:
      properties:
        paste_id:
          type: string
        url:
          type: string
      type: object
...